requirements.txt
numpy
pandas
matplotlib
Pillow
tqdm
kaggle
